+++ libraries/compute-utils/function-views/src/function-view.ts
@@ -854,6 +854,38 @@ export abstract class FunctionView extends DG.ViewBase {
     return this.parentCall?.func.options['mandatoryConsistent'] === 'true';
   }
 
+  protected get hasReadme() {
+    const readmePath = this.func.options['readme'] as string | undefined;
+
+    return !!readmePath;
+  }
+
+  private readmeCache = null as string | null;
+
+  protected async getReadme() {
+    const readmePath = this.func.options['readme'];
+
+    if (!readmePath) return null;
+
+    if (this.readmeCache) return this.readmeCache;
+
+    const packagePath = `System:AppData/${readmePath}`;
+    if (await grok.dapi.files.exists(packagePath)) {
+      const readme = await grok.dapi.files.readAsText(packagePath);
+      this.readmeCache = readme;
+      return readme;
+    }
+
+    const homePath = `${grok.shell.user.name}.home/${readmePath}`;
+    if (await grok.dapi.files.exists(homePath)) {
+      const readme = await grok.dapi.files.readAsText(homePath);
+      this.readmeCache = readme;
+      return readme;
+    }
+
+    return null;
+  }
+
   protected get features(): Record<string, boolean> | string[] {
     return JSON.parse(this.func.options['features'] ?? '{}');
   }