+++ libraries/compute-utils/function-views/src/rich-function-view.ts
@@ -657,6 +657,15 @@ export class RichFunctionView extends FunctionView {
 
     const sensitivityAnalysis = ui.iconFA('analytics', async () => await this.onSALaunch(), 'Run sensitivity analysis');
 
+    const readme = ui.iconFA('info', async () => {
+      if (this.hasReadme) {
+        grok.shell.windows.help.visible = true;
+        // Workaround to deal with help panel bug
+        await new Promise((resolve) => setTimeout(resolve, 100));
+        grok.shell.windows.help.showHelp(ui.markdown((await this.getReadme())!));
+      }
+    });
+
     const newRibbonPanels = [[
       ...super.buildRibbonPanels().flat(),
       ...this.runningOnInput || this.options.isTabbed ? []: [play],
@@ -664,6 +673,7 @@ export class RichFunctionView extends FunctionView {
         ((this.hasUploadMode && this.isUploadMode.value) || (this.isHistoryEnabled && this.runningOnInput))) ? [save] : [],
       ...this.hasUploadMode ? [toggleUploadMode]: [],
       ...this.isSaEnabled ? [sensitivityAnalysis]: [],
+      ...this.hasReadme ? [readme]: [],
     ]];
 
     this.setRibbonPanels(newRibbonPanels);